requirements.txt
# Requirements for AINX ChatBot with Character
transformers
accelerate
bitsandbytes
sentence-transformers
faiss-cpu
langchain
langchain-community
python-dotenv
langchain-huggingface
jupyter
ipywidgets
hf-xet
ddgs
streamlit
torch
torchvision
langchain-core

